fix(BUB-31): added delta time to obstacle, spikes, and rockets

+++ scenes/obstacles/Obstacle.cs
@@ -39,5 +39,5 @@ public partial class Obstacle : CharacterBody2D
     // Called every frame. 'delta' is the elapsed time since the previous frame.
     public override void _Process(double delta) { }
 
-    public virtual void move(double delta) { }
+    public virtual void move(float delta) { }
 }
